\name{DA.MRFA-internal}
\alias{DA.MRFA-internal}
\alias{mrfa_s}
\alias{nrm}
\alias{sr}
\alias{adequacymatrix}
\title{
DA.MRFA-internal
}
\description{
}
\usage{
mrfa_s
nrm
sr
adequacymatrix
}
\arguments{
}
\details{
This functions are for internal use only.
}
\value{
}
\references{
}
\author{
David Navarro-Gonzalez

Urbano Lorenzo-Seva
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ internal }
