\name{DA.MRFA-package}
\alias{DA.MRFA-package}
\title{
Dimensionality Assesment using Minimum Rank Factor Analysis (MRFA)
}
\description{
Package for performing Parallel Analysis using Minimum Rank Factor Analysis (MRFA) . It also include a function to perform the MRFA  only and another function to compute the Greater Lower Bound step for estimating the variables communalities.
}
\usage{
}
\arguments{
}
\details{
For more information about the methods used in each function, please go to each main page.
}
\value{
\item{\code{\link{parallelMRFA}}}{Performs Parallel Analysis using Minimum Rank Factor Analysis (MRFA).}
\item{\code{\link{mrfa}}}{Performs Minimum Rank Factor Analysis (MRFA) procedure.}
\item{\code{\link{GreaterLowerBound}}}{Estimates the communalities of the variables from a factor model.}
}
\references{
}
\author{
David Navarro-Gonzalez

Urbano Lorenzo-Seva
}
\note{
}
\seealso{
}
\examples{
## Each man page contains examples of each function. For a global example use
testme()
}
\keyword{ package }% use one of  RShowDoc("KEYWORDS")
