\name{DA.MRFA-internal}
\alias{DA.MRFA-internal}
\alias{mrfa_s}
\alias{nrm}
\alias{sr}
\alias{adequacymatrix}
\title{
DA.MRFA-internal
}
\description{
This functions are for internal use only.
}
\usage{
mrfa_s
nrm
sr
adequacymatrix
}
\author{
David Navarro-Gonzalez

Urbano Lorenzo-Seva
}
\keyword{ internal }
